Part 4A is OK
I have defined a list: rideHistory， Then, the addVisitorToHistory method was used to add 5 people to the list, and the Boolean checkVisitorFromHistory was used to determine whether that person was in the record based on their name. If they were no longer in the record, it would be false. Then, the length of the list (5 elements, return 5) is obtained by using the size in the numberOfVisitors method. Finally, the for-each method is written to traverse the list, and the output statement is written in the loop inside to output the information of 5 people(It can still get those 5 variables (Person is the parent class)