Part 4B is OK
Add 5 people to the rideHistory list, then iterate through the for each-loop and output information for 5 elements. Then, the rideHistory list is obtained using the sortRideSistory method, and the ages of the 5 individuals are sorted in ascending order using a comparator. Comparing ages one element after another (tourists) . Finally, the output is iterated through an OR each loop, but sorted in ascending order by age