Part 5 is OK
Firstly, I wrote a for loop in class A2, which can loop 11 times. Additionally, I wrote the following code: add 1 person to the queue, loop 11 times to add 11 people to the queue, and then output the information (all of which are integers). Secondly, in the runOneCycle function, let the while loop loop 5 times, and use the poll method to queue the first 5 people in the loop and move them to the rideHistory list. In this way, output the queue and list in A2, where the queue has 6 people and the list (rideHistory) has 5 people.