Part 6 , 7 is OK
Firstly, I import packages for reading and writing files into the Ride class, and write files in the for each loop. The files can be written directly in A2 without any restrictions and can be wrapped at any time. Secondly, the while function is used to determine if there is passenger information in the file. If there is, the information will be read part by part. At the same time, the guest information in the file can be added to the rideHistory list, and the queue can be output in the part7 area. Finally, the file can be read and written  ( I have included try catch exception handling for both methods of reading and writing files in the Ride class, otherwise part 6 and part 7 will report an error and cannot write or read files at the same time )